moved sample_means and related functions back into registration.utils
