+++ suite2p/registration/__init__.py
@@ -1,5 +1,3 @@
-from . import bidiphase, nonrigid, rigid
 from .metrics import get_pc_metrics
-from .utils import sampled_mean
 from .register import register_binary
 from .zalign import compute_zpos
